requirements.txt
django-model-utils==3.1.2
djangorestframework==3.9.0



phonenumbers==8.10.2
django-phonenumber-field==2.3.1


signalwire==1.3



django-datatable==0.3.1
djangorestframework-datatables==0.4.1



channels
dj-database-url
psycopg2
asgi_redis
haikunator
